% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnlme.R
\name{coef.nlme_gastempt}
\alias{coef.nlme_gastempt}
\title{Extract coefficients from nlme_gastempt result}
\usage{
\method{coef}{nlme_gastempt}(object, ...)
}
\arguments{
\item{object}{Result of a call to nlme_gastempt}

\item{...}{other arguments}
}
\value{
a data frame with coefficients. See \code{\link{nlme_gastempt}} for an example.
}
\description{
Extract coefficients from nlme_gastempt result
}
