% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_gastempt.R
\name{coef.stan_gastempt}
\alias{coef.stan_gastempt}
\title{Extract coefficients from stan_gastempt result}
\usage{
\method{coef}{stan_gastempt}(object, ...)
}
\arguments{
\item{object}{Result of a call to stan_gastempt}

\item{...}{other arguments}
}
\value{
a data frame with coefficients. See \code{\link{nlme_gastempt}} for an example.
}
\description{
Extract coefficients from stan_gastempt result
}
