% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnlme.R
\name{plot.nlme_gastempt}
\alias{plot.nlme_gastempt}
\title{Plot data points and fit curve of an nlme_gastempt fit}
\usage{
\method{plot}{nlme_gastempt}(x, ...)
}
\arguments{
\item{x}{Result of a call to nlme_gastempt}

\item{...}{other arguments}
}
\value{
a ggplot object. Use \code{print()} if used non-interactively to show the curve
}
\description{
Plot data points and fit curve of an nlme_gastempt fit
}
