% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_gastempt.R
\name{plot.stan_gastempt}
\alias{plot.stan_gastempt}
\title{Plot data points and fit curve of an stan_gastempt fit}
\usage{
\method{plot}{stan_gastempt}(x, ...)
}
\arguments{
\item{x}{Result of a call to stan_gastempt}

\item{...}{other arguments}
}
\value{
a ggplot object. Use \code{print()} if used
non-interactively to show the curve
}
\description{
Plot data points and fit curve of an stan_gastempt fit
}
