% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Run shiny app demonstrating fit strategies with simulated data}
\usage{
run_shiny()
}
\value{
Not used, starts shiny app
}
\description{
Run shiny app demonstrating fit strategies with simulated data
}
