% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitparams.R
\name{t50}
\alias{t50}
\title{Half-emptying time from nlme parameters}
\usage{
t50(x)
}
\arguments{
\item{x}{Result of a nlme fit, with named components tempt, beta, logbeta,
kappa, logkappa depending on model}
}
\value{
Half-emptying time. Name of evaluated function is returned as attribute \code{fun}. Negative of slope is returned as attribute \code{slope}.
}
\description{
No closed solution known for linexp, we use newton  approximation
}

